refactor(cli): tighten surface and move serialize helper into output stage
Several should-fix items surfaced by independent review:

- Move `to_json_pretty` from `lib.rs` to `stages/output.rs`. Stage 5
  (serialize) now lives entirely in the output stage; `lib.rs` no
  longer pulls in `serde_json` or owns any serialization policy.

- Remove `FileOutcome` and `execute_task` from the public surface.
  `run()` no longer materialises a struct only to destructure it —
  the per-task loop body now inlines `process_file` directly and
  references `task.input_path()` when logging. `ProcessingReport`
  becomes a private struct (it never escaped `run`).

- Introduce `RunSummary::exit_code() -> ExitCode` so the "any errors
  ⇒ FAILURE" policy lives with the summary type rather than in
  `main.rs`. `main.rs` shrinks to a 10-line OS shell that simply
  delegates the exit-code decision.

- Fix a misleading error prefix: `main.rs` previously printed
  "Problem parsing arguments: {err}", but `CliError` now covers
  read/write/serialize failures whose `Display` is already
  informative. Drop the prefix.

- Rename `stages/io.rs` to `stages/files.rs` to remove the cognitive
  collision with `std::io` when both appear in a single file.

- Mark `for_testing` with `#[doc(hidden)]` and an explicit note that
  it is an intra-crate testing hook, reconciling its `pub` visibility
  with the `pub(crate)` visibility of the `domain` / `stages`
  modules it bridges.

- Align stage numbering in `stages/mod.rs` with the 6-stage pipeline
  documented in `lib.rs` (previously drifted to "4a / 4b / 5").

No behavior change; 62 tests pass. No new clippy warnings.
